+++ include/Particle.h
@@ -20,25 +20,25 @@ private:
 	static Vec3d force_field;
 	static Scene *scene;
 
-	static constexpr double density = 2700.; // kg/m3
-	static constexpr double density_medium = 1.; // air kg/m3
-	static constexpr double restitution_coeff = 0.5;
-	static double drag_coefficient; // non-const can not be initialized in the declaration
-	static double uniform_radius;
-	double radius = uniform_radius;
-	double volume() const {
+	static constexpr float density = 2700.; // kg/m3
+	static constexpr float density_medium = 1.; // air kg/m3
+	static constexpr float restitution_coeff = 0.5;
+	static float drag_coefficient; // non-const can not be initialized in the declaration
+	static float uniform_radius;
+	float radius = uniform_radius;
+	float volume() const {
 		return radius * radius * radius * pi * 4.0 / 3.0;
 	}
-	double mass() const {
+	float mass() const {
 		return volume() * density;
 	}
-	double A() const {
+	float A() const {
 		return radius * radius * pi;
 	}
-	double CdA() const {
+	float CdA() const {
 		return drag_coefficient * A();
 	}
-	double CoR() const {
+	float CoR() const {
 		return restitution_coeff;
 	}
 
@@ -46,29 +46,29 @@ private:
 
 public:
 	Particle();
-	Particle(Vec3d _pos, double _r = Particle::uniform_radius) :
+	Particle(Vec3d _pos, float _r = Particle::uniform_radius) :
 			pos(_pos), radius(_r) {
 	}
-	Particle(Vec3d _pos, Vec3d _vel, double _r = Particle::uniform_radius) :
+	Particle(Vec3d _pos, Vec3d _vel, float _r = Particle::uniform_radius) :
 			pos(_pos), vel(_vel), radius(_r) {
 	}
-	Particle(Vec3d _pos, Vec3d _vel, int _ID, double _r =
+	Particle(Vec3d _pos, Vec3d _vel, int _ID, float _r =
 			Particle::uniform_radius) :
 			pos(_pos), vel(_vel), ID(_ID), radius(_r) {
 	}
 	~Particle();
 
-	void advance(double dt);
-	double kineticEnergy();
-	double potentialEnergy();
-	double energy();
+	void advance(float dt);
+	float kineticEnergy();
+	float potentialEnergy();
+	float energy();
 	Vec3d impulse();
 
 	void info();
 	void draw2D();
 	void drawNow2D();
 
-	double distance(const Particle &other) const;
+	float distance(const Particle &other) const;
 
 	void collideToWall(const Boundary &wall);
 	void collideToParticle(Particle &other);
@@ -80,13 +80,13 @@ public:
 	Vec3d overlapVectorWithWall(const Boundary &wall);
 	void size() const;
 
-	void setX(double x) {
+	void setX(float x) {
 		this->pos.x = x;
 	}
 	void setPos(Vec3d pos) {
 		this->pos = pos;
 	}
-	void setR(double r) {
+	void setR(float r) {
 		this->radius = r;
 	}
 	void setV(Vec3d v) {
@@ -96,29 +96,29 @@ public:
 		Particle::scene = scene;
 	}
 
-	static void setCd(const double _drag_coefficient) {
+	static void setCd(const float _drag_coefficient) {
 		drag_coefficient = _drag_coefficient;
 	}
-	static void setUniformRadius(double _uniform_radius) {
+	static void setUniformRadius(float _uniform_radius) {
 		uniform_radius = _uniform_radius;
 	}
 
 	int getID() const {
 		return ID;
 	}
-	double getX() const {
+	float getX() const {
 		return pos.x;
 	}
-	double getY() const {
+	float getY() const {
 		return pos.y;
 	}
-	double getZ() const {
+	float getZ() const {
 		return pos.z;
 	}
-	double getR() const {
+	float getR() const {
 		return radius;
 	}
-	double getM() const {
+	float getM() const {
 		return mass();
 	}
 	Vec3d getV() const {
@@ -127,17 +127,17 @@ public:
 	Vec3d getPos() const {
 		return pos;
 	}
-	double terminalVelocity();
-	double maxFreeFallVelocity(); // in the domain, no drag
-	double maxVelocity();
-	double timeStepLimit();
+	float terminalVelocity();
+	float maxFreeFallVelocity(); // in the domain, no drag
+	float maxVelocity();
+	float timeStepLimit();
 
 	// static getters can not be qualified as const according to the standard
 	// (they do not modify any instance of the class)
-	static double getCd() {
+	static float getCd() {
 		return drag_coefficient;
 	}
-	static double getUniformRadius() {
+	static float getUniformRadius() {
 		return uniform_radius;
 	}
