+++ include/Timer.h
@@ -19,14 +19,14 @@ public:
 		running = false;
 	}
 
-	double milliSeconds() {
+	float milliSeconds() {
 		if (running) {
 			stop();
 		}
 		return std::chrono::duration_cast<std::chrono::microseconds>(
 				end - begin).count() / 1000.;
 	}
-	double seconds() {
+	float seconds() {
 		return milliSeconds() / 1000.;
 	}
 };
